strategy returns void, ioctl cmd is u_long
SCCS-vsn: sys/dev/vn.c 8.8
SCCS-vsn: sys/dev/cd.c 8.3
SCCS-vsn: sys/hp/dev/grf.c 8.5
SCCS-vsn: sys/hp/dev/dca.c 8.3
SCCS-vsn: sys/hp/dev/hil.c 8.3
SCCS-vsn: sys/hp/dev/ite.c 8.3
SCCS-vsn: sys/hp/dev/cons.c 8.3

+++ usr/src/sys/hp/dev/dca.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)dca.c	8.2 (Berkeley) %G%
+ *	@(#)dca.c	8.3 (Berkeley) %G%
  */
 
 #include "dca.h"
@@ -465,7 +465,7 @@ dcamint(unit, dca)
 
 dcaioctl(dev, cmd, data, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 	struct proc *p;
